% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_evap_isotopes.R
\name{add_evap_isotopes}
\alias{add_evap_isotopes}
\title{Add evaporation isotopes to dataset}
\usage{
add_evap_isotopes(tracer, weather, chem_df, start_date, end_date, chem_tracer)
}
\arguments{
\item{tracer}{data frame with water chemistry information for stable isotope.}

\item{weather}{data frame with monthly weather}

\item{chem_df}{data frame with water chemistry information for this analysis,
for getting lake surface temperature.}

\item{start_date}{start date of analysis.}

\item{end_date}{end date of analysis.}

\item{chem_tracer}{name of stable isotope, "d18O" or "d2H".}
}
\value{
Cevap, new data frame with same columns as tracer, but with only
        calculated evaporation values.
}
\description{
Adds evaporation stable isotope measurements, calculated with
"calculate_Cevap" in Cevap_functions.R.
}
