% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_calcification.R
\name{calculate_calcification}
\alias{calculate_calcification}
\title{Calculate calcification rates}
\usage{
calculate_calcification(
  desired_lakes = c("Pleasant", "Long", "Plainfield"),
  water_chem = CSLSdata::water_chem,
  chem_tracer = "d18O",
  start_date = "10-01-2018",
  end_date = "09-30-2019"
)
}
\arguments{
\item{desired_lakes}{vector of lakes to analyze, defaults to
c("Pleasant", "Long", "Plainfield")}

\item{water_chem}{data frame with water chemistry information for all sites.
Defaults to CSLSdata::water_chem.}

\item{chem_tracer}{name of water chemistry parameter to use as a tracer.
Must match the "description" field of CSLSdata::water_chem
for the desired parameter. Defaults to "d18O". Options
include "d18O", "d2H", "CALCIUM TOTAL RECOVERABLE",
"MAGNESIUM TOTAL RECOVERABLE", "CHLORIDE", and
"SODIUM TOTAL RECOVERABLE".}

\item{start_date}{start date of analysis. Defaults to start of WY2019
("2018-10-01").}

\item{end_date}{end date of analysis. Defaults to end of WY2019
("2019-09-30").}
}
\value{
calcify, a data frame with the following columns
}
\description{
Calculates the calcification rate using measurements of calcium, magnesium,
and alkalinity.
}
