% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cevap_functions.R
\name{calculate_Cevap}
\alias{calculate_Cevap}
\title{Evaporation stable isotope}
\usage{
calculate_Cevap(
  atmp,
  ltmp,
  RH,
  Cpcpn,
  Clake,
  chem_tracer,
  month,
  seasonal_correction = TRUE
)
}
\arguments{
\item{atmp}{air temperature (C)}

\item{ltmp}{lake surface temperature (C)}

\item{RH}{relative humidity (percent)}

\item{Cpcpn}{stable isotope composition of precipitation}

\item{Clake}{stable isotope composition of the lake}

\item{chem_tracer}{name of stable isotope, "d18O" or "d2H"}

\item{month}{numeric value of month (e.g., 1, 2, 3,...12)}

\item{seasonal_correction}{logical defaults to TRUE to incorporate seasonal
correction factors for d18O_atm based on
difference between calculated and reported
d18O_atm values in Krabbenhoft et al. (1990).}
}
\value{
Cevap - the isotope composition of evaporation
}
\description{
Calculates the stable isotope composition of evaporation based on equation 5
in Krabbenhoft et al. (1990). Note that alpha* is equivalent to alpha^-1.
}
\references{
Krabbenhoft, D. P., C. J. Bowser, M. P. Anderson, and J. W.
  Valley. (1990). Estimating Groundwater Exchange with Lakes: 1. The Stable
  Isotope Mass Balance Method. Water Resources Research, 26(10):2445-2453.
  https://doi.org/10.1029/WR026i010p02445
}
