% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chem_bal.R
\name{calculate_chem_bal}
\alias{calculate_chem_bal}
\title{Filter SWIMS by parameter}
\usage{
calculate_chem_bal(
  lake,
  start_date = as_datetime(mdy("10-01-2018")),
  end_date = as_datetime(mdy("09-30-2019"))
)
}
\arguments{
\item{lake}{- lakes of interest, e.g., c("Pleasant", "Long", "Plainfield")}

\item{start_date}{start date of analysis. Defaults to
as_datetime(mdy("10-01-2018")) but can be NULL to include
entier water budget dataset.}

\item{end_date}{end date of analysis. Defaults to
as_datetime(mdy("09-30-2019")) but can be NULL to include
entier water budget dataset.}
}
\value{
chem_bal - a data frame with the date, parameter (e.g., Ca, Mg), result (concentration), flux (e.g., precip, GWin, lake), volume (m^3), and mass balance (mg/L*m^3.
}
\description{
Subsets SWIMS dataset for a CSLS lake to only measurements for a parameter of
interest. Labels with lake, and defines measurements as lake, upgradient
well, or downgradient well based on the previous month of groundwater and
lake levels.
}
