% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_level_diff.R
\name{calculate_level_diff}
\alias{calculate_level_diff}
\title{Calculate difference between groundwater and lake levels}
\usage{
calculate_level_diff(lakes = c("Pleasant", "Long", "Plainfield"))
}
\arguments{
\item{lakes}{vector with lakes of interest, defaults to c("Pleasant", "Long",
"Plainfield")}
}
\value{
water_levels, a data frame with the following columns:
\item{date}{date of observation, POSIXct}
\item{site_id}{site id of the groundwater well}
\item{diff_m}{difference between groundwater levels and lake levels (m)}
}
\description{
For given list of lakes, calculates the difference between groundwater levels
and lake levels (positive = groundwater is higher).
}
