% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_solute_balance.R
\name{calculate_solute_balance}
\alias{calculate_solute_balance}
\title{Calculate solute mass balance}
\usage{
calculate_solute_balance(
  parameters = c("CALCIUM TOTAL RECOVERABLE", "CONDUCTIVITY, UMHOS/CM @ 25C",
    "CHLORIDE", "MAGNESIUM TOTAL RECOVERABLE", "NITROGEN NH3-N DISS",
    "NITROGEN NO3+NO2 DISS (AS N)", "NITROGEN TOTAL", "PH LAB",
    "POTASSIUM TOTAL RECOVERABLE", "SODIUM TOTAL RECOVERABLE", "SULFATE TOTAL"),
  start_date = as_datetime("2018-10-01"),
  end_date = as_datetime("2019-09-30")
)
}
\arguments{
\item{parameters}{names of water chemsitry parameters to analyze mass balance
for. Defaults to c("CALCIUM TOTAL RECOVERABLE",
"CONDUCTIVITY, UMHOS/CM @ 25C", "CHLORIDE", "MAGNESIUM
TOTAL RECOVERABLE", "NITROGEN NH3-N DISS", "NITROGEN
NO3+NO2 DISS (AS N)", "NITROGEN TOTAL", "PH LAB",
"POTASSIUM TOTAL RECOVERABLE", "SODIUM TOTAL RECOVERABLE",
"SULFATE TOTAL").}

\item{start_date}{start date of analysis. Defaults to start of WY2019
("2018-10-01").}

\item{end_date}{end date of analysis. Defaults to end of WY2019
("2019-09-30").}
}
\value{
mass_fluxes
}
\description{
Calculates the solute mass balance.
}
