% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_solute_dynamics.R
\name{calculate_solute_dynamics}
\alias{calculate_solute_dynamics}
\title{Calculate solute mass balance}
\usage{
calculate_solute_dynamics(
  start_date = as_datetime("2018-10-01"),
  end_date = as_datetime("2019-09-30"),
  method = "load",
  lakes = c("Pleasant", "Long", "Plainfield"),
  annual = FALSE,
  parameter = "CALCIUM TOTAL RECOVERABLE"
)
}
\arguments{
\item{start_date}{start date of analysis. Defaults to start of WY2019
("2018-10-01").}

\item{end_date}{end date of analysis. Defaults to end of WY2019
("2019-09-30").}

\item{method}{"load" or "calc", defaults to "load".}

\item{lakes}{name of all lakes in CSLS in the order of desired levels.
Defaults to c("Pleasant", "Long", "Plainfield")}

\item{annual}{logical defaults to "FALSE" to return daily values, switch to
"TRUE" for annual summary of uptake (Ca) or
sedimentation/release and GW inflow (TP).}

\item{parameter}{description of parameter to use dynamic lake model on. Can
be either "CALCIUM TOTAL RECOVERABLE" or "PHOSPHORUS TOTAL".}
}
\value{
mass_fluxes
}
\description{
Calculates the solute mass balance.
}
