% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_water_balance.R
\name{calculate_water_balance}
\alias{calculate_water_balance}
\title{Calculate water balance}
\usage{
calculate_water_balance(
  parameter = "d18O",
  start_date = as_datetime("2018-08-22"),
  end_date = as_datetime("2019-11-05"),
  dt = "day",
  method = "indirect_load"
)
}
\arguments{
\item{parameter}{parameter to use for water balance calculations, defaults to "d18O".}

\item{start_date}{start date of calculations}

\item{end_date}{end date of calculations}

\item{dt}{desired final time step ("day", "month", or "annual")}

\item{method}{approach for calculating water balance, either "direct",
"indirect_calc", or "indirect_load". Defaults to
"indirect_load".}
}
\value{
lake_fluxes
}
\description{
Calculates the water balance using a stable isotope tracer using either a) a
direct approach, where GWin is directly calculated from measured and
interpolated values, or b) an indirect approach, where a dynamic lake model
with a daily time step is set up and GW residence time is solved for using
optimization.
}
