% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_daily_gradients.R
\name{classify_daily_gradients}
\alias{classify_daily_gradients}
\title{Calculate groundwater gradients relative to lakes}
\usage{
classify_daily_gradients(
  lakes = c("Pleasant", "Long", "Plainfield"),
  threshold = 0.01,
  duration = 30
)
}
\arguments{
\item{lakes}{lakes of interest, e.g., c("Pleasant", "Long", "Plainfield")}

\item{threshold}{minimum difference (m) between water levels in order to
assess gradient, defaults to 0.01m (i.e., the precision of
HOBO water level loggers)}

\item{duration}{window (days) to use for calculating median difference
between groundwater and lake levels, defaults to 30 days.}
}
\value{
gw_gradients, a data frame with the following columns:
\item{date}{date of record, POSIXct}
\item{site_id}{site id of groundwater well}
\item{gradient}{whether well should be considered "upgradient",
                "downgradient", or unknown ("") on that day}
}
\description{
This function determines whether a well is upgradient, downgradient, or
unknown at all recorded dates for all wells associated with the lakes of
interest. Uses the given duration and threshold values to classify wells
(median difference over given duration must be greater in magnitude than the
threshold value).
}
