% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_SWIMS_results.R
\name{classify_SWIMS_results}
\alias{classify_SWIMS_results}
\title{Classify SWIMS results}
\usage{
classify_SWIMS_results(
  lakes = c("Pleasant", "Long", "Plainfield"),
  precip = "Hancock",
  threshold = 0.01,
  duration = 30
)
}
\arguments{
\item{lakes}{lakes of interest, e.g., c("Pleasant", "Long", "Plainfield")}

\item{precip}{name of precip site, defaults to "Hancock".}

\item{threshold}{minimum difference (m) between water levels in order to
assess gradient, defaults to 0.01m (i.e., the precision of
HOBO water level loggers)}

\item{duration}{window (days) to use for calculating median difference
between groundwater and lake levels, defaults to 30 days.}
}
\value{
SWIMS, a data frame with all columns included in the CSLSdata frame
}
\description{
This function first classifies each well site by daily lake-gw gradients,
then classifies SWIMS samples results by site and site type (lake, upgradient
well, downgradient well, deep well, or precipitation)
}
