% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidate_values.R
\name{consolidate_values}
\alias{consolidate_values}
\title{Consolidate vectors with missing values}
\usage{
consolidate_values(col1, col2, missing = "")
}
\arguments{
\item{col1}{one vector}

\item{col2}{second vector}

\item{missing}{what a "missing" value looks like, defaults to "". Could also
be NA, NULL, etc.}
}
\value{
col - a vector with consolidated values
}
\description{
Given two vectors, each of which may have some missing values, consolidate
into one vector with the minimum number of missing values.
}
