% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_gradient.R
\name{define_gradient}
\alias{define_gradient}
\title{Define groundwater gradient}
\usage{
define_gradient(values, threshold = 0.01)
}
\arguments{
\item{value}{mean difference between groundwater and lake levels (positive =
groundwater is higher)}

\item{thrshold}{value to use to evaluate threshold, defaults to 0.01 (i.e.,
the precision of HOBO water level loggers)}
}
\value{
gradient - a string noting gradient (upgradient, downgradient, or "")
}
\description{
Given a value and a threshold, define whether groundwater is upgradient,
downgradient, or undetermined.
}
