% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_lake_wells.R
\name{define_lake_wells}
\alias{define_lake_wells}
\title{Define upgradient and downgradient wells}
\usage{
define_lake_wells(lake, SWIMS, static_gw = FALSE, threshold = 0.01)
}
\arguments{
\item{lake}{- lake of interest from the CSLS (i.e., "Pleasant", "Plainfield",
or "Long")}

\item{SWIMS}{- data frame with SWIMS info}

\item{static_gw}{switch defaults to FALSE to calculate groundwater gradients based on lake and groundwater levels. If TRUE, uses static definitions of gradient in the dictionary.}

\item{threshold}{- defaults to 0.01 (m) to indicate that the difference in
lake levels and groundwater levels over the previous month
should be greater than 1cm in order to classify a well as
upgradient.}
}
\value{
SWIMS - same as inputs SWIMS data frame, but with a column for
                "site_type" which is either "lake", "upgradient" or
                "downgradient".
}
\description{
This function takes an inputs SWIMS data frame and classifies each
measurement as associated with the "lake", an "upgradent" well, or a
"downgradient" well based on the previous 30 days of lake and groundwater
measurements.
}
