% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lake_calcium.R
\name{dynamic_lake_calcium}
\alias{dynamic_lake_calcium}
\title{Dynamic lake model - Calcium}
\usage{
dynamic_lake_calcium(k0, df)
}
\arguments{
\item{k0}{zero-order rate constant for Calcium uptake.}

\item{df}{data frame with lake inputs for lake of interest.}
}
\value{
df, a data frame with dynamic lake model for calcium
}
\description{
Calculates lake concentration of calcium using dynamic lake model approach
}
