% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lake_phosphorus.R
\name{dynamic_lake_phosphorus}
\alias{dynamic_lake_phosphorus}
\title{Dynamic lake model - Phosphorus}
\usage{
dynamic_lake_phosphorus(par, df, deposition = 5.75)
}
\arguments{
\item{par}{vector with initial sediment concentration, etc.}

\item{df}{data frame with lake inputs for lake of interest.}

\item{deposition}{estimated wet/dry deposition of phosphorus (kg/km2-yr)}
}
\value{
df, a data frame with dynamic lake model for phosphorus
}
\description{
Calculates lake concentration of phosphorus using dynamic lake model approach
}
