% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lake_water_balance.R
\name{dynamic_lake_water_balance}
\alias{dynamic_lake_water_balance}
\title{Optimize water balance}
\usage{
dynamic_lake_water_balance(t_res, t_switch, df, lake, parameter)
}
\arguments{
\item{t_res}{vector with residence times for first time period (t_res[1]) and
second time period (t_res[2]).}

\item{t_switch}{date at which second time period begins}

\item{df}{daily inputs for water balance calculations incl. preciptation (vol
& concentration), evaporation (vol & concentration), ice formation
(vol & concentration), and lake volume.}

\item{lake}{current lake to evaluate}

\item{parameter}{parameter used for calculating water balance, usually "d18O"}
}
\value{
RMSE, root mean square error of match with lake measurements of
        parameter.
}
\description{
Calculates RMSE of measured lake concentrations with those calculated with
the dynamic lake model (daily timestep) for use in optimization procedure.
Assumes use of two time periods for groundwater inflow calculcations (two
periods may have different residence times).
}
