% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_lst.R
\name{filter_lst}
\alias{filter_lst}
\title{Filter to lake surface temperature}
\usage{
filter_lst(chem_df = CSLSdata::water_chem, use_HOBO = TRUE)
}
\arguments{
\item{chem_df}{data frame with water chemistry information for all sites.
Defaults to CSLSdata::water_chem.}

\item{use_HOBO}{logical defaults to TRUE to use HOBO data for temperature
measurements. if false, uses field profile data.}
}
\value{
lst, same data frame as chem_df, but subset to only lake surface
        temperature measurements.
}
\description{
Filters large dataset of water chemistry parameters to only lake surface
temperature measurements.
}
