% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_parameter.R
\name{filter_parameter}
\alias{filter_parameter}
\title{Filter SWIMS by parameter}
\usage{
filter_parameter(lakes = c("Pleasant", "Long", "Plainfield"), parameter,
  plotting_name, static_gw = FALSE)
}
\arguments{
\item{lakes}{- lakes of interest, e.g., c("Pleasant", "Long", "Plainfield")}

\item{parameter}{- parameter to subset by, must exactly match DNR description
(e.g., "ALUMINUM,TOTAL RECOVERABLE")}

\item{plotting_name}{- name of parameter to use for plotting, e.g., "Total
Recoverable Alumninum". Units will be added in based
on SWIMS field.}

\item{static_gw}{- defaults to FALSE to use well levels and lake levels to
determine upgradient vs. downgradient wells. If TRUE, uses
static definition from dictionary.}
}
\value{
parameter - a data frame with SWIMS data subset to only parameter of
                    interest for each lake, with site type (upgradient,
                    downgradient, or lake) and lake (Pleasant, Plainfield, or
                    Long), noted.
}
\description{
Subsets SWIMS dataset for a CSLS lake to only measurements for a parameter of
interest. Labels with lake, and defines measurements as lake, upgradient
well, or downgradient well based on the previous month of groundwater and
lake levels.
}
