% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ice_on_off.R
\name{find_ice_on_off}
\alias{find_ice_on_off}
\title{Ice On and Off Dates}
\usage{
find_ice_on_off(lst, ice_tmp = 4)
}
\arguments{
\item{lst}{a data frame with date and ltmp of lake surface temperature
measurements.}

\item{ice_tmp}{temperature below which assume ice on, default to 4 deg C}
}
\value{
ice, a list with "ice_on" and "ice_off" dates.
}
\description{
Estimates ice on and ice off based on lake surface temperature. Assumes "ice
on" is the first day between Sept and May that surface water temperature goes
below 4 degres C and "ice off" is the last day between Sept and May that
surface water temperature is below 4 degrees C.
}
