% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_sensitivity.R
\name{gw_sensitivity}
\alias{gw_sensitivity}
\title{Filter SWIMS by parameter}
\usage{
gw_sensitivity(lake, GW_factor)
}
\arguments{
\item{lake}{- lakes of interest, e.g., c("Pleasant", "Long", "Plainfield")}
}
\value{
chem_bal - a data frame with the date, parameter (e.g., Ca, Mg), result (concentration), flux (e.g., precip, GWin, lake), volume (m^3), and mass balance (mg/L*m^3.
}
\description{
Subsets SWIMS dataset for a CSLS lake to only measurements for a parameter of
interest. Labels with lake, and defines measurements as lake, upgradient
well, or downgradient well based on the previous month of groundwater and
lake levels.
}
