% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hancock.R
\docType{data}
\name{hancock}
\alias{hancock}
\title{Daily Hancock air temperature and precipitation}
\format{
A data frame with the following columns:
\describe{
 \item{date}{date of observation}
 \item{atmp_min}{minimum air temperature observed this day (deg C)}
 \item{atmp_max}{maximum air temperature observed this day (deg C)}
 \item{pcpn}{daily precipitation (mm)}
 }
}
\usage{
data(hancock)
}
\description{
Daily Hancock air temperature and precipitation
}
\keyword{datasets}
