% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_chem_values.R
\name{interpolate_chem_values}
\alias{interpolate_chem_values}
\title{Interpolate water chemistry values}
\usage{
interpolate_chem_values(df, dt = "month")
}
\arguments{
\item{df}{data frame with water chemistry information for tracer/solute.
        Includes:
* **date:** date of chem measurement
* **lake:** lake associated with measurement
* **site_type:** type of site (e.g., lake, precipitaiton, upgradient)
* **result:** value of measurement}

\item{dt}{unit of time to use for interpolation. Defaults to "month", can
also be "day".}
}
\value{
df, a data frame with interpolated values.
}
\description{
Interpolates water chemistry values for consistent comparisons across site
types. Can interpolate to a daily time step or a monthly time step. If
monthly, assumes end-of-month values are desired.
}
