% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_tracer_values.R
\name{interpolate_tracer_values}
\alias{interpolate_tracer_values}
\title{Interpolate tracer values}
\usage{
interpolate_tracer_values(tracer)
}
\arguments{
\item{tracer}{data frame with water chemistry information for tracer}
}
\value{
tracer, a data frame with interpolated end-of-month tracer values.
}
\description{
Interpolates tracer values to the last day of each month for consistent
comparisons across site types.
}
