% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_dx.R
\docType{data}
\name{K_dx}
\alias{K_dx}
\title{Ksat and distance to lake for wells}
\format{
A data frame with the following columns:
\describe{
 \item{site_id}{uniqe id associated with each well}
 \item{K_ft_d}{saturated hydraulic conductivity (ft/d)}
 \item{dx_ft}{distance from well to lake (ft)}
 }
}
\usage{
data(K_dx)
}
\description{
Ksat and distance to lake for wells
}
\keyword{datasets}
