% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lillie_mason_lakes.R
\docType{data}
\name{lillie_mason_lakes}
\alias{lillie_mason_lakes}
\title{Data from Wisconsin Lakes survey by Lillie and Mason}
\format{
A data frame with the following columns:
\describe{
 \item{location}{WI, central, or seepage}
 \item{description}{parameter, corresponds to dnr_description where possible}
 \item{units}{units of measured value}
 \item{n}{number of lakes sampled}
 \item{mean}{mean value}
 \item{std}{standard deviation of values}
 \item{median}{median value}
 \item{min}{minimum value}
 \item{max}{maximum value}
 }
}
\usage{
data(lillie_mason_lakes)
}
\description{
Data from Wisconsin Lakes survey by Lillie and Mason
}
\references{
Lillie, R.A. and J.W. Mason 1983. Limnological Characteristics of Wisconsin Lakes. Technical Bulletin No. 138. Wisconsin Department of Natural Resources. Madison, Wisconsin.
}
\keyword{datasets}
