% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgTOmeq.R
\docType{data}
\name{mgTOmeq}
\alias{mgTOmeq}
\title{mg/L to meq/L conversion table}
\format{
A data frame with the following columns:
\describe{
 \item{description}{dnr parameter description}
 \item{mgTOmeq}{multiply value in mg/L by this value to yield meq/L}
 \item{mgTOmmol}{multiply value in mg/L by this value to yield molarity}
 }
}
\usage{
data(mgTOmeq)
}
\description{
mg/L to meq/L conversion table
}
\references{
Schwartz, F.W. and H. Zhang. 2003. Fundamentals of Groundwater. John Wiley & Sons, New York, NY. Table 16.4.
}
\keyword{datasets}
