% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_solutes.R
\name{optimize_solutes}
\alias{optimize_solutes}
\title{Optimize dynamic lake model}
\usage{
optimize_solutes(par, df, parameter)
}
\arguments{
\item{par}{vector with tuning parameters. k0 for Calcium. P_sed, ks, ks_tmp,
kr, and kr_tmp for TP.}

\item{df}{data frame with lake inputs for lake of interest.}

\item{parameter}{description of parameter to use dynamic lake model on. Can
be either "CALCIUM TOTAL RECOVERABLE" or "PHOSPHORUS TOTAL".}
}
\value{
RMSE, root mean square error of match with lake measurements of
        parameter.
}
\description{
Calculates RMSE of measured lake concentrations with those calculated with
the dynamic lake model (daily timestep) for use in optimization procedure.
}
