% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_names.R
\docType{data}
\name{param_names}
\alias{param_names}
\title{Parameter Names}
\format{A data frame:
\describe{
  \item{NADP_param}{name of NADP precipitation parameters}
  \item{CSLS_param}{name of CSLS lake and well parameters from SWIMS}
  \item{CSLS_names}{name of CSLS parameters to use for plotting labels}
}}
\usage{
data(param_names)
}
\description{
Maps the names of parameters in NADP precipitation dataset and CSLS wells and
lakes SWIMS dataset to one another.
}
\keyword{datasets}
