% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chem_functions.R
\name{plot_bulk}
\alias{plot_bulk}
\title{Parameter plot}
\usage{
plot_bulk(SWIMS_parameter, text_size = 12)
}
\arguments{
\item{SWIMS_parameter}{subsetted version of SWIMS data for parameter of interest.}

\item{text_size}{size of text on plots, defaults to 12pt.}
}
\value{
plot_obj, the plot object created by ggplot.
}
\description{
Plots Wells vs. Lake measurements for a parameter of interest at all lakes.
}
