% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparison.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Plot comparison of lake water chemistry}
\usage{
plot_comparison(
  description,
  x_title = description,
  text_size = 12,
  no_seepage = TRUE
)
}
\arguments{
\item{description}{dnr description of parameter to plot}

\item{x_title}{x-axis title to use on plot, defaults to description.}

\item{text_size}{size of text on plots, defaults to 12pt.}

\item{no_seepage}{logical defaults to TRUE to filter out all WI seepage lakes
from comparison.}
}
\value{
plot_obj, the plot object created by ggplot.
}
\description{
Plots a comparison of lake water chemistry for the CSLS study lakes with
values for all of Wisconsin, Central Wisconsin, and Wisconsin seepage lakes.
}
