% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chem_functions.R
\name{plot_inputs}
\alias{plot_inputs}
\title{Plot Chem Mass Balance}
\usage{
plot_inputs(chem_bal, lake, as_pcnt = TRUE, text_size = 12)
}
\arguments{
\item{chem_bal}{a data frame with the chemistry mass balance for multiple
parameters}

\item{lake}{lake of interest (i.e., "Pleasant", "Long", or "Plainfield")}

\item{as_pcnt}{logical defaults to FALSE to plot as masses, if true plots as
percent of incoming mass.}

\item{text_size}{defaults to 12 point text size.}
}
\value{
plot_obj - a plot object with the chem mass balance
}
\description{
This function creates a plot object to show the chemistry mass balance of
multiple parameters
}
