% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chem_functions.R
\name{plot_levels}
\alias{plot_levels}
\title{Plot Levels}
\usage{
plot_levels(water_level_diff, SWIMS_gw)
}
\arguments{
\item{water_level_diff}{a data frame with the date, site_no, obs_type,
level_m (of gw levels), and diff_m between daily gw level
observations and lake level observations.}

\item{SWIMS_gw}{a data frame with the date and site_id of all SWIMS gw measurmements.}
}
\value{
plot_obj - a plot object with water level difference and dates of
                   isotope measurements for each lake site.
}
\description{
This function creates a plot object to compare the difference in lake level
and well level measurements for all sites at a given lake, paired with when
isotope measurements were collected and what the d18O results were.
}
