% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mass_in.R
\name{plot_mass_in}
\alias{plot_mass_in}
\title{Plot Chem Mass Balance}
\usage{
plot_mass_in(chem_bal, as_pcnt = TRUE, text_size = 12)
}
\arguments{
\item{chem_bal}{a data frame with the chemistry mass balance for multiple
parameters}

\item{as_pcnt}{logical defaults to FALSE to plot as masses, if true plots as
percent of incoming mass.}

\item{text_size}{defaults to 12 point text size.}
}
\value{
plot_obj - a plot object with the chem mass balance
}
\description{
This function creates a plot object to show the chemistry mass balance of
multiple parameters
}
