% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries_by_year.R
\name{plot_timeseries_by_year}
\alias{plot_timeseries_by_year}
\title{Parameter plot}
\usage{
plot_timeseries_by_year(df, yaxis_type = "number", text_size = 12)
}
\arguments{
\item{df}{subsetted version of SWIMS data for parameter of interest.}

\item{yaxis_type}{defaults to "number" to use continuous y axis. Can also be
"log" for semi-log plot.}

\item{text_size}{size of text on plots, defaults to 12pt.}

\item{site_types}{possible site types to plot}

\item{site_labels}{how to display names of site types in legends}

\item{site_colors}{colors to use for site types}

\item{site_shapes}{shapes to use for site types}
}
\value{
plot_obj, the plot object created by ggplot.
}
\description{
Plots Wells vs. Lake measurements for a parameter of interest at all lakes.
}
