% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_balance_inputs.R
\name{process_balance_inputs}
\alias{process_balance_inputs}
\title{Calculate water balance}
\usage{
process_balance_inputs(
  parameter = "d18O",
  start_date = as_datetime("2018-10-01"),
  end_date = as_datetime("2019-09-30"),
  dt = "day",
  no_ice = TRUE,
  C_evap = 0,
  C_ice = 0
)
}
\arguments{
\item{parameter}{name of water chemistry parameter to grab concentration
information for. Must match the "description" field of
CSLSdata::water_chem for the desired parameter. Defaults to
"d18O". Options include "d18O", "d2H", "CALCIUM TOTAL
RECOVERABLE", "MAGNESIUM TOTAL RECOVERABLE", "CHLORIDE", and
"SODIUM TOTAL RECOVERABLE".}

\item{start_date}{start date of analysis. Defaults to start of WY2019
("2018-10-01").}

\item{end_date}{end date of analysis. Defaults to end of WY2019
("2019-09-30").}

\item{dt}{desired time step of inputs (e.g., "day" or "month")}

\item{no_ice}{logical defaults to TRUE to ignore ice formation.}

\item{C_evap}{concentraion in evaporation, defaults to zero}

\item{C_ice}{concentration in ice, defaults to zero.}
}
\value{
lake_inputs
}
\description{
Calculates the water balance using desired tracer, e.g., stable isotopes or
conservative anion/cation.
}
