% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWIMS.R
\docType{data}
\name{SWIMS}
\alias{SWIMS}
\title{SWIMS sample results}
\format{
A data frame:
\describe{
  \item{lake}{Hancock, Pleasant, Long, or Plainfield}
  \item{site_type}{lake, upgradient, downgradient, deep, or precipitation}
  \item{date}{datetime of sample}
  \item{dnr_parameter}{DNR parameter id}
  \item{description}{description of parameter analyzed (e.g., ALKALINITY
                     TOTAL CACO3)}
  \item{result}{measured result (chracter)}
  \item{units}{unites of measured result, e.g. "DEGREES C"}
  \item{depth1_m}{depth at which sample was collected (m)}
  \item{depth2_m}{if a depth range, other end of depths at which sample was
                  collected (m)}
  \item{site_id}{unique id for CSLS site, e.g., "LL-01"}
  \item{lod}{limit of detection}
  \item{loq}{limit of quality}
  \item{flag}{flag for sample quality, inlcudes "AGE", "BLANK", "COMMENT",
              "DUPLICATE", "LOD", "LOQ", or "NONE"}
  \item{flag_reason}{detailed explanation for flag, if not "NONE"}
}
}
\usage{
data(SWIMS)
}
\description{
Samples from the SWIMS database, classified by CSLS site type.
}
\keyword{datasets}
