% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_water_balance.R
\name{calculate_water_balance}
\alias{calculate_water_balance}
\title{Calculate water balance}
\usage{
calculate_water_balance(
  desired_lakes = c("Pleasant", "Long", "Plainfield"),
  chem_df = CSLSdata::water_chem,
  chem_tracer = "MAGNESIUM TOTAL RECOVERABLE",
  start_date = "10-01-2018",
  end_date = "09-30-2019",
  annual = FALSE,
  no_ice = FALSE,
  ice_on = "12-01-2018",
  ice_off = "03-31-2019",
  use_kniffin = TRUE
)
}
\arguments{
\item{desired_lakes}{vector of lakes to analyze, defaults to
c("Pleasant", "Long", "Plainfield")}

\item{chem_df}{data frame with water chemistry information for all sites.
Defaults to CSLSdata::water_chem.}

\item{chem_tracer}{name of water chemistry parameter to use as a tracer.
Must match the "description" field of CSLSdata::water_chem
for the desired parameter. Defaults to "d18O". Options
include "d18O", "d2H", "CALCIUM TOTAL RECOVERABLE",
"MAGNESIUM TOTAL RECOVERABLE", "CHLORIDE", and
"SODIUM TOTAL RECOVERABLE".}

\item{start_date}{start date of analysis. Defaults to start of WY2019
("2018-10-01").}

\item{end_date}{end date of analysis. Defaults to end of WY2019
("2019-09-30").}

\item{annual}{logical defaults to FALSE to calculate water balance on a
monthly time step. If TRUE, calculates for entire timeseries
(start_date to end_date), assumed to be one year.}

\item{no_ice}{logical defaults to FALSE. If TRUE, calculates water balance
for ice-free months only.}

\item{ice_on}{date of ice on, defaults to "12-01-2018"}

\item{ice_off}{date of ice off, defaults to "03-31-2019"}

\item{use_kniffin}{logical defaults to TRUE to add in Kniffin precipitation
isotope samples for months without CSLS precipitation
isotope samples}
}
\value{
fluxes
}
\description{
Calculates the water balance using desired tracer, e.g., stable isotopes or
conservative anion/cation.
}
