% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_parameter.R
\name{filter_parameter}
\alias{filter_parameter}
\title{Filter SWIMS by parameter}
\usage{
filter_parameter(
  SWIMS,
  parameter,
  plotting_name,
  no_blanks = TRUE,
  no_dups = TRUE,
  no_age = FALSE
)
}
\arguments{
\item{SWIMS}{data frame with SWIMS information for lakes of interest}

\item{parameter}{- parameter to subset by, must exactly match DNR description
(e.g., "ALUMINUM,TOTAL RECOVERABLE")}

\item{plotting_name}{- name of parameter to use for plotting, e.g., "Total
Recoverable Alumninum". Units will be added in based
on SWIMS field.}

\item{no_blanks}{logical defaults to TRUE to exclude blank samples}

\item{no_dups}{logical defaults to TRUE to exclude duplicate samples}

\item{no_age}{logical defaults to FALSE to include samples that were analyzed
past the holding date.}
}
\value{
SWIMS_parameter, same data frame, with only the results for the
        parameter of interest. Also includes the given plotting name ("name").
}
\description{
Subsets SWIMS dataset for a CSLS lake to only measurements for a parameter of
interest.
}
