% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Parameter plot}
\usage{
plot_summary(
  df,
  yaxis_type = "number",
  text_size = 12,
  site_types = c("upgradient", "nogradient", "downgradient", "lake", "lake_bottom"),
  site_labels = c("Upgr.", "Nogr.", "Downgr.", "Lake Surf", "Lake Bot."),
  site_colors = c("#33A02C", "#B15928", "#B2DF8A", "#FB9A99", "#E31A1C")
)
}
\arguments{
\item{df}{subsetted version of SWIMS data for parameter of interest.}

\item{yaxis_type}{defaults to "number" to use continuous y axis. Can also be
"log" for semi-log plot.}

\item{text_size}{size of text on plots, defaults to 12pt.}

\item{site_types}{possible site types to plot}

\item{site_labels}{how to display names of site types in legends}

\item{site_colors}{colors to use for site types}
}
\value{
plot_obj, the plot object created by ggplot.
}
\description{
Plots Wells vs. Lake measurements for a parameter of interest at all lakes.
}
